@@ -1,5 +1,4 @@
 import 'package:flutter/animation.dart';
-import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_circular_chart/src/animated_circular_chart.dart';
 import 'package:flutter_circular_chart/src/entry.dart';
@@ -17,22 +16,22 @@ class CircularChart {
 
   final List<CircularChartStack> stacks;
   final CircularChartType chartType;
-  final SegmentEdgeStyle edgeStyle;
+  final SegmentEdgeStyle? edgeStyle;
 
-  factory CircularChart.empty({@required CircularChartType chartType}) {
+  factory CircularChart.empty({required CircularChartType chartType}) {
     return new CircularChart(<CircularChartStack>[], chartType);
   }
 
   factory CircularChart.fromData({
-    @required Size size,
-    @required List<CircularStackEntry> data,
-    @required CircularChartType chartType,
-    @required bool percentageValues,
-    @required double startAngle,
-    Map<String, int> stackRanks,
-    Map<String, int> entryRanks,
-    double holeRadius,
-    SegmentEdgeStyle edgeStyle,
+    required Size size,
+    required List<CircularStackEntry> data,
+    required CircularChartType chartType,
+    required bool percentageValues,
+    required double startAngle,
+    Map<String?, int>? stackRanks,
+    Map<String?, int>? entryRanks,
+    double? holeRadius,
+    SegmentEdgeStyle? edgeStyle,
   }) {
     final double _holeRadius = holeRadius ?? size.width / (2 + data.length);
     final double stackDistance =
@@ -43,14 +42,14 @@ class CircularChart {
     List<CircularChartStack> stacks = new List<CircularChartStack>.generate(
       data.length,
       (i) => new CircularChartStack.fromData(
-            stackRanks[data[i].rankKey] ?? i,
-            data[i].entries,
-            entryRanks,
-            percentageValues,
-            startRadius + i * stackDistance,
-            stackWidth,
-            startAngle,
-          ),
+        stackRanks![data[i].rankKey] ?? i,
+        data[i].entries,
+        entryRanks,
+        percentageValues,
+        startRadius + i * stackDistance,
+        stackWidth,
+        startAngle,
+      ),
     );
 
     return new CircularChart(stacks, chartType, edgeStyle: edgeStyle);
@@ -68,7 +67,7 @@ class CircularChartTween extends Tween<CircularChart> {
   @override
   CircularChart lerp(double t) => new CircularChart(
         _stacksTween.lerp(t),
-        begin.chartType,
-        edgeStyle: end.edgeStyle,
+        begin!.chartType,
+        edgeStyle: end!.edgeStyle,
       );
 }
